requirements.txt
matplotlib>=3.8.0
numpy>=1.24.0
pandas>=2.0.0
seaborn>=0.12.0
plotly>=5.15.0
openpyxl>=3.1.0

# 圖表優化增強功能
kaleido>=0.2.1
pillow>=10.0.0
scipy>=1.11.0

# 可選的進階視覺化套件
# bokeh>=3.2.0
# holoviews>=1.17.0
# plotnine>=0.12.0
